Move plotting helper function from mlutils.models to mlutils.plots
